requirements.txt
fastapi
uvicorn
requests
pydantic
pdfplumber
pillow
pytesseract
openai
python-dotenv
